+++ DatadogRUM/Tests/Instrumentation/WatchdogTerminations/WatchdogTerminationCheckerTests.swift
@@ -10,25 +10,17 @@ import DatadogInternal
 import TestUtilities
 
 final class WatchdogTerminationCheckerTests: XCTestCase {
-    // swiftlint:disable implicitly_unwrapped_optional
-    var sut: WatchdogTerminationChecker!
-    // swiftlint:enable implicitly_unwrapped_optional
+    var sut: WatchdogTerminationChecker = .init(appStateManager: .mockRandom(), featureScope: FeatureScopeMock())
 
     func testNoPreviousState_NoWatchdogTermination() throws {
-        given(isSimulator: .mockRandom())
-
-        XCTAssertFalse(sut.isWatchdogTermination(launch: .mockRandom(), from: nil, to: .mockRandom()))
+        XCTAssertFalse(sut.isWatchdogTermination(launch: .mockRandom(), deviceInfo: .mockWith(isSimulator: false), from: nil, to: .mockRandom()))
     }
 
     func testIsSimulatorBuild_NoWatchdogTermination() throws {
-        given(isSimulator: true)
-
-        XCTAssertFalse(sut.isWatchdogTermination(launch: .mockRandom(), from: .mockRandom(), to: .mockRandom()))
+        XCTAssertFalse(sut.isWatchdogTermination(launch: .mockRandom(), deviceInfo: .mockWith(isSimulator: false), from: .mockRandom(), to: .mockRandom()))
     }
 
     func testIsDebugging_NoWatchdogTermination() throws {
-        given(isSimulator: false)
-
         let previous = WatchdogTerminationAppState(
             appVersion: .mockAny(),
             osVersion: .mockAny(),
@@ -37,14 +29,13 @@ final class WatchdogTerminationCheckerTests: XCTestCase {
             wasTerminated: .mockAny(),
             isActive: .mockAny(),
             vendorId: .mockAny(),
-            processId: .mockAny()
+            processId: .mockAny(),
+            trackingConsent: .mockRandom()
         )
-        XCTAssertFalse(sut.isWatchdogTermination(launch: .mockRandom(), from: previous, to: .mockRandom()))
+        XCTAssertFalse(sut.isWatchdogTermination(launch: .mockRandom(), deviceInfo: .mockWith(isSimulator: false), from: previous, to: .mockRandom()))
     }
 
     func testDifferentAppVersions_NoWatchdogTermination() throws {
-        given(isSimulator: false)
-
         let previous = WatchdogTerminationAppState(
             appVersion: "1.0.0",
             osVersion: .mockAny(),
@@ -53,7 +44,8 @@ final class WatchdogTerminationCheckerTests: XCTestCase {
             wasTerminated: .mockAny(),
             isActive: .mockAny(),
             vendorId: .mockAny(),
-            processId: .mockAny()
+            processId: .mockAny(),
+            trackingConsent: .mockRandom()
         )
 
         let current = WatchdogTerminationAppState(
@@ -64,15 +56,14 @@ final class WatchdogTerminationCheckerTests: XCTestCase {
             wasTerminated: .mockAny(),
             isActive: .mockAny(),
             vendorId: .mockAny(),
-            processId: .mockAny()
+            processId: .mockAny(),
+            trackingConsent: .mockRandom()
         )
 
-        XCTAssertFalse(sut.isWatchdogTermination(launch: .mockRandom(), from: previous, to: current))
+        XCTAssertFalse(sut.isWatchdogTermination(launch: .mockRandom(), deviceInfo: .mockWith(isSimulator: false), from: previous, to: current))
     }
 
     func testApplicationDidCrash_NoWatchdogTermination() throws {
-        given(isSimulator: false)
-
         let previous = WatchdogTerminationAppState(
             appVersion: .mockAny(),
             osVersion: .mockAny(),
@@ -81,7 +72,8 @@ final class WatchdogTerminationCheckerTests: XCTestCase {
             wasTerminated: .mockAny(),
             isActive: .mockAny(),
             vendorId: .mockAny(),
-            processId: .mockAny()
+            processId: .mockAny(),
+            trackingConsent: .mockRandom()
         )
 
         let current = WatchdogTerminationAppState(
@@ -92,15 +84,14 @@ final class WatchdogTerminationCheckerTests: XCTestCase {
             wasTerminated: .mockAny(),
             isActive: .mockAny(),
             vendorId: .mockAny(),
-            processId: .mockAny()
+            processId: .mockAny(),
+            trackingConsent: .mockRandom()
         )
 
-        XCTAssertFalse(sut.isWatchdogTermination(launch: .init(didCrash: true), from: previous, to: current))
+        XCTAssertFalse(sut.isWatchdogTermination(launch: .init(didCrash: true), deviceInfo: .mockWith(isSimulator: false), from: previous, to: current))
     }
 
     func testApplicationWasTerminated_NoWatchdogTermination() throws {
-        given(isSimulator: false)
-
         let previous = WatchdogTerminationAppState(
             appVersion: "1.0.0",
             osVersion: .mockAny(),
@@ -109,7 +100,8 @@ final class WatchdogTerminationCheckerTests: XCTestCase {
             wasTerminated: true,
             isActive: .mockAny(),
             vendorId: .mockAny(),
-            processId: .mockAny()
+            processId: .mockAny(),
+            trackingConsent: .mockRandom()
         )
 
         let current = WatchdogTerminationAppState(
@@ -120,15 +112,14 @@ final class WatchdogTerminationCheckerTests: XCTestCase {
             wasTerminated: .mockAny(),
             isActive: .mockAny(),
             vendorId: .mockAny(),
-            processId: .mockAny()
+            processId: .mockAny(),
+            trackingConsent: .mockRandom()
         )
 
-        XCTAssertFalse(sut.isWatchdogTermination(launch: .init(didCrash: false), from: previous, to: current))
+        XCTAssertFalse(sut.isWatchdogTermination(launch: .init(didCrash: false), deviceInfo: .mockWith(isSimulator: false), from: previous, to: current))
     }
 
     func testDifferentOSVersions_NoWatchdogTermination() throws {
-        given(isSimulator: false)
-
         let previous = WatchdogTerminationAppState(
             appVersion: "1.0.0",
             osVersion: "1.0.0",
@@ -137,7 +128,8 @@ final class WatchdogTerminationCheckerTests: XCTestCase {
             wasTerminated: .mockAny(),
             isActive: .mockAny(),
             vendorId: .mockAny(),
-            processId: .mockAny()
+            processId: .mockAny(),
+            trackingConsent: .mockRandom()
         )
 
         let current = WatchdogTerminationAppState(
@@ -148,15 +140,14 @@ final class WatchdogTerminationCheckerTests: XCTestCase {
             wasTerminated: .mockAny(),
             isActive: .mockAny(),
             vendorId: .mockAny(),
-            processId: .mockAny()
+            processId: .mockAny(),
+            trackingConsent: .mockRandom()
         )
 
-        XCTAssertFalse(sut.isWatchdogTermination(launch: .init(didCrash: false), from: previous, to: current))
+        XCTAssertFalse(sut.isWatchdogTermination(launch: .init(didCrash: false), deviceInfo: .mockWith(isSimulator: false), from: previous, to: current))
     }
 
     func testDifferentBootTimes_NoWatchdogTermination() throws {
-        given(isSimulator: false)
-
         let previous = WatchdogTerminationAppState(
             appVersion: "1.0.0",
             osVersion: "1.0.0",
@@ -165,7 +156,8 @@ final class WatchdogTerminationCheckerTests: XCTestCase {
             wasTerminated: .mockAny(),
             isActive: .mockAny(),
             vendorId: .mockAny(),
-            processId: .mockAny()
+            processId: .mockAny(),
+            trackingConsent: .mockRandom()
         )
 
         let current = WatchdogTerminationAppState(
@@ -176,15 +168,14 @@ final class WatchdogTerminationCheckerTests: XCTestCase {
             wasTerminated: .mockAny(),
             isActive: .mockAny(),
             vendorId: .mockAny(),
-            processId: .mockAny()
+            processId: .mockAny(),
+            trackingConsent: .mockRandom()
         )
 
-        XCTAssertFalse(sut.isWatchdogTermination(launch: .init(didCrash: false), from: previous, to: current))
+        XCTAssertFalse(sut.isWatchdogTermination(launch: .init(didCrash: false), deviceInfo: .mockWith(isSimulator: false), from: previous, to: current))
     }
 
     func testDifferentVendorId_NoWatchdogTermination() throws {
-        given(isSimulator: false)
-
         let previous = WatchdogTerminationAppState(
             appVersion: "1.0.0",
             osVersion: "1.0.0",
@@ -193,7 +184,8 @@ final class WatchdogTerminationCheckerTests: XCTestCase {
             wasTerminated: .mockAny(),
             isActive: .mockAny(),
             vendorId: "foo",
-            processId: .mockAny()
+            processId: .mockAny(),
+            trackingConsent: .mockRandom()
         )
 
         let current = WatchdogTerminationAppState(
@@ -204,15 +196,14 @@ final class WatchdogTerminationCheckerTests: XCTestCase {
             wasTerminated: .mockAny(),
             isActive: .mockAny(),
             vendorId: "bar",
-            processId: .mockAny()
+            processId: .mockAny(),
+            trackingConsent: .mockRandom()
         )
 
-        XCTAssertFalse(sut.isWatchdogTermination(launch: .init(didCrash: false), from: previous, to: current))
+        XCTAssertFalse(sut.isWatchdogTermination(launch: .init(didCrash: false), deviceInfo: .mockWith(isSimulator: false), from: previous, to: current))
     }
 
     func testSDKWasStoppedAndStarted_NoWatchdogTermination() throws {
-        given(isSimulator: false)
-
         let pid = UUID()
 
         let previous = WatchdogTerminationAppState(
@@ -223,7 +214,8 @@ final class WatchdogTerminationCheckerTests: XCTestCase {
             wasTerminated: .mockAny(),
             isActive: .mockAny(),
             vendorId: "foo",
-            processId: pid
+            processId: pid,
+            trackingConsent: .mockRandom()
         )
 
         let current = WatchdogTerminationAppState(
@@ -234,15 +226,14 @@ final class WatchdogTerminationCheckerTests: XCTestCase {
             wasTerminated: .mockAny(),
             isActive: .mockAny(),
             vendorId: "foo",
-            processId: pid
+            processId: pid,
+            trackingConsent: .mockRandom()
         )
 
-        XCTAssertFalse(sut.isWatchdogTermination(launch: .init(didCrash: false), from: previous, to: current))
+        XCTAssertFalse(sut.isWatchdogTermination(launch: .init(didCrash: false), deviceInfo: .mockWith(isSimulator: false), from: previous, to: current))
     }
 
     func testApplicationWasInBackground_NoWatchdogTermination() throws {
-        given(isSimulator: false)
-
         let previous = WatchdogTerminationAppState(
             appVersion: "1.0.0",
             osVersion: "1.0.0",
@@ -251,7 +242,8 @@ final class WatchdogTerminationCheckerTests: XCTestCase {
             wasTerminated: .mockAny(),
             isActive: false,
             vendorId: "foo",
-            processId: .mockAny()
+            processId: .mockAny(),
+            trackingConsent: .mockRandom()
         )
 
         let current = WatchdogTerminationAppState(
@@ -262,15 +254,14 @@ final class WatchdogTerminationCheckerTests: XCTestCase {
             wasTerminated: .mockAny(),
             isActive: .mockAny(),
             vendorId: "foo",
-            processId: .mockAny()
+            processId: .mockAny(),
+            trackingConsent: .mockRandom()
         )
 
-        XCTAssertFalse(sut.isWatchdogTermination(launch: .init(didCrash: false), from: previous, to: current))
+        XCTAssertFalse(sut.isWatchdogTermination(launch: .init(didCrash: false), deviceInfo: .mockWith(isSimulator: false), from: previous, to: current))
     }
 
     func testApplicationWasInForeground_WatchdogTermination() throws {
-        given(isSimulator: false)
-
         let previous = WatchdogTerminationAppState(
             appVersion: "1.0.0",
             osVersion: "1.0.0",
@@ -279,7 +270,8 @@ final class WatchdogTerminationCheckerTests: XCTestCase {
             wasTerminated: .mockAny(),
             isActive: true,
             vendorId: "foo",
-            processId: UUID()
+            processId: UUID(),
+            trackingConsent: .mockRandom()
         )
 
         let current = WatchdogTerminationAppState(
@@ -290,31 +282,10 @@ final class WatchdogTerminationCheckerTests: XCTestCase {
             wasTerminated: .mockAny(),
             isActive: true,
             vendorId: "foo",
-            processId: UUID()
+            processId: UUID(),
+            trackingConsent: .mockRandom()
         )
 
-        XCTAssertTrue(sut.isWatchdogTermination(launch: .init(didCrash: false), from: previous, to: current))
-    }
-
-    // MARK: Helpers
-
-    func given(isSimulator: Bool) {
-        let deviceInfo: DeviceInfo = .init(
-            name: .mockAny(),
-            model: .mockAny(),
-            osName: .mockAny(),
-            osVersion: .mockAny(),
-            osBuildNumber: .mockAny(),
-            architecture: .mockAny(),
-            isSimulator: isSimulator,
-            vendorId: .mockAny(),
-            isDebugging: .mockAny(),
-            systemBootTime: .mockAny()
-        )
-
-        sut = WatchdogTerminationChecker(
-            appStateManager: .mockRandom(),
-            deviceInfo: deviceInfo
-        )
+        XCTAssertTrue(sut.isWatchdogTermination(launch: .init(didCrash: false), deviceInfo: .mockWith(isSimulator: false), from: previous, to: current))
     }
 }